part4 4.21

+++ controllers/blogs.js
@@ -4,14 +4,7 @@ import jwt from "jsonwebtoken";
 import {SECRET} from '../utils/config.js';
 import Blog from '../models/blog.js';
 import User from '../models/user.js';
-
-// const getTokenFrom = request => {
-//   const authorization = request.get('Authorization');
-//   if(authorization && authorization.startsWith('Bearer ')){
-//     return authorization.replace('Bearer ', ''); 
-//   }  
-//   return null;
-// }
+import { infoM } from '../utils/logger.js';
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1});
@@ -29,9 +22,8 @@ blogsRouter.get('/:id', async (request, response) => {
   
 blogsRouter.post('/', async (request, response) => {
     const body = request.body;
-
     const decodedToken = jwt.verify(request.body.token, SECRET);
-    console.log(decodedToken);
+
     if(!decodedToken.id){
       return response.status(401).json({error: 'invalid token'});
     }
@@ -46,8 +38,16 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const deleteBlog = await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blogToDelete = await Blog.findById(request.params.id);
+  const decodedToken = jwt.verify(request.body.token, SECRET);
+  infoM(decodedToken);
+  
+  if(decodedToken.id === blogToDelete.user.toString()){
+    const deleteBlog = await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }else{
+    return response.status(401).json({error: 'Unauthorized'});
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
